+++ src/ui/rand.rs
@@ -1,9 +1,9 @@
 use rand::{self, random}; //Random stuff
-pub fn rand_round(input: f64) -> u128 {
+pub fn rand_round(input: f64) -> u64 {
     if input % 1.0 < random() {
-        input.floor() as u128
+        input.floor() as u64
     } else {
-        input.ceil() as u128
+        input.ceil() as u64
     }
 } //Rounds stuff randomly: eg
   //11.15 has a 15% chance of being 12 and an 85% chance of being 11
